Refactor admin ID type from uint8 to uint64 across models, repositories, and services for consistency and improved handling of larger ID values

+++ app/middleware/auth.go
@@ -104,7 +104,7 @@ func (m *appMiddleware) Auth() gin.HandlerFunc {
 
 				if newUser.Role == models.RoleAdmin {
 					if err = m.repository.CreateAdminAvailability(&models.AdminAvailability{
-						AdminID: uint8(newUser.ID),
+						AdminID: newUser.ID,
 					}); err != nil {
 						c.AbortWithStatusJSON(http.StatusInternalServerError, helpers.NewResponse(http.StatusInternalServerError, "Failed to create admin availability", nil, nil))
 						return
