+++ app/repositories/admin_availability.go
@@ -19,13 +19,13 @@ func (r *appRepository) GetAdminAvailabilityByAdminID() (*models.AdminAvailabili
 	return &availability, nil
 }
 
-func (r *appRepository) IncrementAdminConversationCount(adminID uint8) error {
+func (r *appRepository) IncrementAdminConversationCount(adminID uint64) error {
 	return r.Conn.Model(&models.AdminAvailability{}).
 		Where("admin_id = ?", adminID).
 		UpdateColumn("current_conversations", gorm.Expr("current_conversations + ?", 1)).Error
 }
 
-func (r *appRepository) DecrementAdminConversationCount(adminID uint8) error {
+func (r *appRepository) DecrementAdminConversationCount(adminID uint64) error {
 	return r.Conn.Model(&models.AdminAvailability{}).
 		Where("admin_id = ? AND current_conversations > 0", adminID).
 		UpdateColumn("current_conversations", gorm.Expr("current_conversations - ?", 1)).Error
